🚚 api: move error to src/team context
