+++ api/src/shared/application/error-manager.js
@@ -229,9 +229,6 @@ function _mapToHttpError(error) {
   if (error instanceof DomainErrors.ApplicationWithInvalidClientSecretError) {
     return new HttpErrors.UnauthorizedError('The client secret is invalid.');
   }
-  if (error instanceof DomainErrors.OrganizationArchivedError) {
-    return new HttpErrors.UnprocessableEntityError(error.message);
-  }
   if (error instanceof DomainErrors.MissingAttributesError) {
     return new HttpErrors.UnprocessableEntityError(error.message);
   }
