+++ api/src/team/domain/usecases/create-membership.usecase.js
@@ -1,5 +1,5 @@
-import { OrganizationArchivedError } from '../../../shared/domain/errors.js';
 import { roles } from '../../../shared/domain/models/Membership.js';
+import { OrganizationArchivedError } from '../errors.js';
 
 const createMembership = async function ({ userId, organizationId, membershipRepository, organizationRepository }) {
   const organization = await organizationRepository.get(organizationId);
