+++ api/src/team/domain/usecases/create-organization-invitations.usecase.js
@@ -1,6 +1,6 @@
 import bluebird from 'bluebird';
 
-import { OrganizationArchivedError } from '../../../../src/shared/domain/errors.js';
+import { OrganizationArchivedError } from '../errors.js';
 
 const createOrganizationInvitations = async function ({
   organizationId,
