+++ api/tests/integration/application/error-manager_test.js
@@ -659,14 +659,6 @@ describe('Integration | API | Controller Error', function () {
       expect(responseDetail(response)).to.equal('Authentication method already exists.');
     });
 
-    it('responds UnprocessableEntityError when an OrganizationArchived error occurs', async function () {
-      routeHandler.throws(new DomainErrors.OrganizationArchivedError());
-      const response = await server.requestObject(request);
-
-      expect(response.statusCode).to.equal(UNPROCESSABLE_ENTITY_ERROR);
-      expect(responseDetail(response)).to.equal("L'organisation est archivée.");
-    });
-
     it('responds UnprocessableEntityError when an MissingAttributes error occurs', async function () {
       routeHandler.throws(new DomainErrors.MissingAttributesError());
       const response = await server.requestObject(request);
