+++ api/tests/team/unit/application/http-error-mapper-configuration.test.js
@@ -2,6 +2,7 @@ import { HttpErrors } from '../../../../src/shared/application/http-errors.js';
 import { teamDomainErrorMappingConfiguration } from '../../../../src/team/application/http-error-mapper-configuration.js';
 import {
   AlreadyExistingAdminMemberError,
+  OrganizationArchivedError,
   UncancellableOrganizationInvitationError,
 } from '../../../../src/team/domain/errors.js';
 import { expect } from '../../../test-helper.js';
@@ -31,5 +32,19 @@ describe('Unit | Team | Application | HttpErrorMapperConfiguration', function ()
 
     //then
     expect(error).to.be.instanceOf(HttpErrors.UnprocessableEntityError);
+  });
+
+  it('instantiates UnprocessableEntityError when OrganizationArchivedError', async function () {
+    //given
+    const httpErrorMapper = teamDomainErrorMappingConfiguration.find(
+      (httpErrorMapper) => httpErrorMapper.name === OrganizationArchivedError.name,
+    );
+
+    //when
+    const error = httpErrorMapper.httpErrorFn(new OrganizationArchivedError());
+
+    //then
+    expect(error).to.be.instanceOf(HttpErrors.UnprocessableEntityError);
+    expect(error.message).to.equal("L'organisation est archivée.");
   });
 });
