+++ api/tests/team/unit/domain/usecases/create-membership.usecase.test.js
@@ -1,5 +1,5 @@
-import { OrganizationArchivedError } from '../../../../../src/shared/domain/errors.js';
 import { Membership } from '../../../../../src/shared/domain/models/index.js';
+import { OrganizationArchivedError } from '../../../../../src/team/domain/errors.js';
 import { createMembership } from '../../../../../src/team/domain/usecases/create-membership.usecase.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../../test-helper.js';
