+++ api/tests/team/unit/domain/usecases/create-organization-invitation-by-admin.usecase.test.js
@@ -1,5 +1,5 @@
-import { OrganizationArchivedError } from '../../../../../src/shared/domain/errors.js';
 import { Membership } from '../../../../../src/shared/domain/models/index.js';
+import { OrganizationArchivedError } from '../../../../../src/team/domain/errors.js';
 import { createOrganizationInvitationByAdmin } from '../../../../../src/team/domain/usecases/create-organization-invitation-by-admin.usecase.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../../test-helper.js';
